abstracting the connection to mysql away from app.ts
